v3 of methods firm
