--- Balancer.rb
@@ -1,28 +1,0 @@
-#Esta clase administra los threads que buscan noticias nuevas
-class Balancer
-
-  #empieza los threads del pool
-  def startThreads
-  end
-
-  #procesa una estrategia
-  def process(startegy)
-  end
-
-  #anade una estrategia a la cola
-  def queue(strategy)
-  end
-
-  #verifica el estado de un thread
-  def checkThreadStatus(thread)
-  end
-
-  #reinicia un thread
-  def restartThread(thread)
-  end
-
-  #monitorea todos los threads
-  def monitorThreads
-  end
-
-end
