--- DatabaseAdapter.rb
@@ -1,12 +1,0 @@
-#Esta es la clase que se utiliza para insertar/editar/eliminar
-#noticias a la base de datos
-
-require "pg"
-
-class DatabaseAdapter
-
-  #Este metodo inserta nuevas noticias a la base de datos
-  def save(news)
-  end
-
-end
