--- Loader.rb
@@ -1,9 +1,0 @@
-#Esta clase activa los patrones estrategias que buscan
-#nuevas noticias
-class Loader
-
-  #Esta clase carga un patron estrategia
-  def load(pattern)
-  end
-
-end
