--- LogDatabaseAdapter.rb
@@ -1,16 +1,0 @@
-#Esta clase se utiliza para insertar la ultima noticia de una fuente
-#de esta forma nos aseguramos de no tener noticias repetidas
-
-require "sqlite3"
-
-class LogDatabseAdapter
-
-  #Este metodo inserta nuevas noticias a la base de datos
-  def save(news)
-  end
-
-  #revisa la base de datos para ver si contiene la noticia
-  def contains
-  end
-
-end
