--- ProviderStrategy.rb
@@ -1,10 +1,0 @@
-#Esta es la clase base para todas las estrategias
-#buscadoras de noticias
-class ProviderStrategy
-
-  #este metodo recibe la ultima noticia que se busco de la fuente
-  #y retorna una lista de todas las noticias nuevas
-  def run(last)
-  end
-
-end
