+++ balancer.rb
@@ -1,0 +1,29 @@
+# Manaages the threads that gets fresh news
+
+class Balancer
+
+  # starts the thread pool
+  def startThreads
+  end
+
+  # process a given strategy
+  def process(startegy)
+  end
+
+  # adds a strategy to the queue
+  def queue(strategy)
+  end
+
+  # verifies the current state of a thread
+  def checkThreadStatus(thread)
+  end
+
+  # restarts a given thread
+  def restartThread(thread)
+  end
+
+  # checks all threads
+  def monitorThreads
+  end
+
+end
