+++ database/database_adapter.rb
@@ -1,0 +1,11 @@
+# Adapter for the database
+
+require "pg"
+
+class DatabaseAdapter
+
+  # saves a news to the database. Returns a bool
+  def save(news)
+  end
+
+end
