+++ loader.rb
@@ -1,0 +1,9 @@
+# Loads the strategy patterns that fetch fresh news.
+
+class Loader
+
+  # loads a given strategy
+  def load(pattern)
+  end
+
+end
