+++ provider_strategy.rb
@@ -1,0 +1,10 @@
+# Base class for all strategies.
+
+class ProviderStrategy
+
+ # Recieves the last news fetch from the source
+ # and returns a list of new news
+  def run(last)
+  end
+
+end
